+++ netplotbrain/plotting/plot_cm.py
@@ -12,7 +12,7 @@ def _plot_connectivitymatrix(ax, edges, nodes=None, node_color=None, node_colorb
 
     # Dataframe to array
     if isinstance(nodes, pd.DataFrame):
-        number_of_nodes=len(nodes)+1
+        number_of_nodes=len(nodes)
     else:
         number_of_nodes=edges[['i', 'j']].max().max()+1
 
@@ -29,13 +29,22 @@ def _plot_connectivitymatrix(ax, edges, nodes=None, node_color=None, node_colorb
     #    cm_boundarycolor = 'black'
     #else:
     #    cm_boundarycolor = None
+    # Set cm_order to node_colorby if it is set to auto and node_color is set
+    if cm_order == 'auto' and node_colorby is not None:
+        cm_order = node_colorby
+        if cm_boundary == 'auto': 
+            cm_boundary = True
+    elif cm_order == 'auto':
+        cm_order = None
     if cm_order is not None:
-        if cm_order == 'auto' and node_colorby is not None:
-            cm_order = node_colorby
-            if cm_boundary == 'auto': 
-                cm_boundary = True
-        communities = nodes[cm_order].astype('category').cat.codes.values
-        nodeorder = np.argsort(communities)
+        # Check if cm_order consists of unqiue indicies or are communities/groups
+        if len(np.unique(cm_order))==len(cm_order):
+            nodeorder = cm_order
+        else:
+            communities = nodes[cm_order].astype('category').cat.codes.values
+            nodeorder = np.argsort(communities)
+    else:
+        nodeorder = np.arange(number_of_nodes)
     # If it is still set to auto, set ot None
     if cm_boundary == 'auto':
         cm_boundary = None
@@ -60,6 +69,11 @@ def _plot_connectivitymatrix(ax, edges, nodes=None, node_color=None, node_colorb
 
     # Plot
     # Add squares for different communities to the connectivity plot
+    if cm_border:            
+        ax.plot([0, 0], [0, number_of_nodes], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
+        ax.plot([number_of_nodes, 0], [number_of_nodes, number_of_nodes], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
+        ax.plot([0, number_of_nodes], [0, 0], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
+        ax.plot([number_of_nodes, number_of_nodes], [0, number_of_nodes], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
     if cm_boundary:
         for coms in np.unique(communities):
             no = np.where(communities[nodeorder] == coms)[0]
@@ -67,8 +81,3 @@ def _plot_connectivitymatrix(ax, edges, nodes=None, node_color=None, node_colorb
             ax.plot([no[-1]+1, no[0]], [no[-1]+1, no[-1]+1], color=node_color[nodeorder[no[0]]], linewidth=cm_boundarywidth, transform=tr,zorder=10)
             ax.plot([no[0], no[-1]+1], [no[0], no[0]], color=node_color[nodeorder[no[0]]], linewidth=cm_boundarywidth, transform=tr,zorder=10)
             ax.plot([no[-1]+1, no[-1]+1], [no[0], no[-1]+1], color=node_color[nodeorder[no[0]]], linewidth=cm_boundarywidth, transform=tr,zorder=10)
-    if cm_border:            
-        ax.plot([0, 0], [0, len(nodes)], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
-        ax.plot([len(nodes), 0], [len(nodes), len(nodes)], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
-        ax.plot([0, len(nodes)], [0, 0], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=5)
-        ax.plot([len(nodes), len(nodes)], [0, len(nodes)], color=cm_bordercolor, linewidth=cm_borderwidth, transform=tr,zorder=150)